+++ NEWS.md
@@ -4,6 +4,7 @@
 
 * Fixed issue with bibliographies containing 'other-ids' (#174).
 * Fixed encoding issues with `bibliography_entries()` on Windows (#158).
+* Add fallback entry formatter as tibble
 
 # vitae 0.4.2